Made generating more user friendly and more efficient
Now instead of checking for errors from the user, generating the base url and then rechecking all the values to add their url segment. It checks for errors and adds all segments with there value, since if you added pram to the url buts it blank it ignores it.

Added past builds to gitgnore
Expanded the todo list