Added 2 more test
Added 2 more test, test  5 works fine. test 6 kinda works the problem with it is the api is not applying the tournament parameter.
Test 4 was failing because the a couple events got deleted from the api and test 1 for season summary 
was failing because the number of teams went from 3213 to 4577