Event listing database got changed again and made testing os undependent
going to fix the issue now and comment the test out to wait until they stop
list of changes.
team number for event listings now returns this.
Object reference not set to an instance of an object. 
changes all district codes to null
added back the removed events
district code for event listing now returns this.
Object reference not set to an instance of an object
Even with another season trying to use any arguments are returning
Object reference not set to an instance of an object.